+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Changed
+- **Breaking**: Changed browser event namespace from `ember-nexus-` to `ember-nexus-sdk-`, closes #83.
 
 ## 0.0.58 - 2024-07-08
 ### Changed